fix: duplicate chat, create for BtnChat component
