+++ src/components/BtnChat/BtnChat.module.scss
@@ -1,0 +1,13 @@
+.wrapperChat {
+  position: sticky;
+  bottom: 30px;
+}
+
+.chatBtn {
+  position: relative;
+  display: block;
+  margin-left: auto;
+  height: 85px;
+  width: 85px;
+  border-radius: 100%;
+}
