+++ src/pages/Game/Game.module.scss
@@ -2,20 +2,6 @@
   margin-bottom: 50px;
 }
 
-.wrapperChat {
-  position: sticky;
-  bottom: 30px;
-}
-
-.chatBtn {
-  position: relative;
-  display: block;
-  margin-left: auto;
-  height: 85px;
-  width: 85px;
-  border-radius: 100%;
-}
-
 .gameInner {
   display: flex;
   justify-content: space-between;
