+++ src/pages/Lobby/Lobby.module.scss
@@ -13,18 +13,4 @@
     display: flex;
     align-items: center;
   }
-
-  .wrapperChat {
-    position: sticky;
-    bottom: 30px;
-  }
-
-  .chatBtn {
-    position: relative;
-    display: block;
-    margin-left: auto;
-    height: 85px;
-    width: 85px;
-    border-radius: 100%;
-  }
 }
